requirements.txt
certifi==2017.11.5
decorator==4.1.2
dj-database-url==0.4.2
Django==2.0
future==0.16.0
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.0
parso==0.1.0
pexpect==4.3.0
pickleshare==0.7.4
pkg-resources==0.0.0
prompt-toolkit==1.0.15
psycopg2==2.7.3.2
ptyprocess==0.5.2
Pygments==2.2.0
python-decouple==3.1
python-telegram-bot==8.1.1
pytz==2017.3
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
wcwidth==0.1.7
